move FunctionSelector into eth/ pkg
