+++ core/store/models/service_agreement_test.go
@@ -5,6 +5,7 @@ import (
 	"strings"
 	"testing"
 
+	"chainlink/core/eth"
 	"chainlink/core/internal/cltest"
 	"chainlink/core/store/assets"
 	"chainlink/core/store/models"
@@ -184,8 +185,8 @@ func TestEncumbrance_ABI(t *testing.T) {
 
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
-			fs := func(s string) models.FunctionSelector {
-				return models.BytesToFunctionSelector(hexutil.MustDecode(s))
+			fs := func(s string) eth.FunctionSelector {
+				return eth.BytesToFunctionSelector(hexutil.MustDecode(s))
 			}
 			enc := models.Encumbrance{
 				Payment:                test.payment,
